Create patient_tracker app, and move Admission to it
