+++ patient_tracker/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class PatientTrackerConfig(AppConfig):
+    name = 'patient_tracker'
